Changed type of vars in 15

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,7 +8,7 @@
 
 void fibonacci(void)
 {
-	long long int a, b, c, sum;
+	long int a, b, c, sum;
 
     c = 2;
 	a = 1;
